Retrieve & return any fulltext links from response
The EDS 'search' response contains a FullText element which exposes any
links to fulltext versions of the record. This commit makes use of this
to allow us to return an array of fulltext links with each record in our
response.

Includes:

- Refactor Api.pm to move create authentication functions that can be
used by both 'search' and 'retrieve'
- Add retrieval of any CustomLink elements that may be
present in the EDS API response
- Add retrieval of any Link elements that may be present in the EDS API
response
- Add retrieval method when dealing with Link elements, as detailed in
"Important note" here:
https://connect.ebsco.com/s/article/EBSCO-Discovery-Service-API-User-Guide
- Collate links gathered and return them in the response