Modify search string values as per EBSCO advice
PTFS ticket #43032
A customer reported DOI searching not working. It turns out that the
search values must be URL escaped. However, based on prior advice from
EBSCO we are manually escaping (with a \) certain characters, a .
character being one of them. So the . characters in a DOI were being
escaped thusly \. - but when I added the URL escaping the slash was
being escaped to %5C which was causing the search to return no results.

This commit removes the . character entirely, in addition to the other
characters we are already removing - it also adds URL escaping of the
entire remaining value